chat_app/requirements.txt
aioredis==1.3.1
asgiref==2.3.2
astroid==2.3.3
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
cffi==1.14.4
Django==2.2.17
channels==2.0.2
colorama==0.4.1
constantly==15.1.0
cryptography==3.3.1
daphne==2.2.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.0
pylint==2.4.4
pyOpenSSL==20.0.1
pytz==2020.5
service-identity==18.1.0
six==1.13.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.0
wrapt==1.11.2
zope.interface==5.2.0
